+++ library/libsoft1/libflow/libnet/3.app/net.rdp.c
@@ -2,7 +2,7 @@
 #define u16 unsigned short
 #define u32 unsigned int
 #define u64 unsigned long long
-int serve_rdp(u64 fd, u64 type, u8* buf, u64 len)
+u64 serve_rdp(u64 fd, u64 type, u8* buf, u64 len)
 {
 	return 0;
 }
